refactor(sc-hub): split connection helpers under file-size cap
