+++ crates/bacnet-transport/src/sc_hub.rs
@@ -24,17 +24,19 @@ use tokio_tungstenite::tungstenite::Message;
 use tokio_tungstenite::WebSocketStream;
 use tracing::{debug, warn};
 
+mod connection;
+mod helpers;
 mod relay;
+
+use connection::accept_loop;
+use helpers::*;
 
 use relay::{
     build_hub_relay_message, hub_relay_recipient_vmacs, hub_relay_target, HubRelayReject,
     HubRelayTarget,
 };
 
-use crate::sc_frame::{
-    decode_sc_message, encode_sc_message, ScFunction, ScMessage, Vmac, BACNET_SC_HUB_SUBPROTOCOL,
-    BROADCAST_VMAC,
-};
+use crate::sc_frame::{decode_sc_message, encode_sc_message, ScFunction, ScMessage, Vmac};
 
 type TlsStream = tokio_rustls::server::TlsStream<tokio::net::TcpStream>;
 type WsSink = SplitSink<WebSocketStream<TlsStream>, Message>;
@@ -170,161 +172,6 @@ impl ScHub {
             task.abort();
             let _ = task.await;
         }
-    }
-}
-
-// ---------------------------------------------------------------------------
-// Accept loop
-// ---------------------------------------------------------------------------
-
-// Closure passed to `accept_hdr_async` returns the upstream tungstenite
-// `ErrorResponse`, whose size is fixed by the library. The clippy lint can't
-// be addressed without changing the foreign signature.
-#[allow(clippy::result_large_err)]
-async fn accept_loop(
-    listener: TcpListener,
-    tls_acceptor: TlsAcceptor,
-    hub_vmac: Vmac,
-    hub_uuid: DeviceUuid,
-    clients: Clients,
-) {
-    // Track active TCP connections (pre-handshake) to limit DoS surface.
-    let active_connections = std::sync::Arc::new(std::sync::atomic::AtomicUsize::new(0));
-    const MAX_ACTIVE_CONNECTIONS: usize = 512;
-
-    // Heartbeat sweep: periodically check for idle clients and send HeartbeatRequest.
-    // Per Annex AB.6.3, peers initiate heartbeats to detect idle/dead connections.
-    const HEARTBEAT_CHECK_INTERVAL_SECS: u64 = 30;
-    const HEARTBEAT_IDLE_THRESHOLD_SECS: u64 = 60;
-    {
-        let clients_for_hb = clients.clone();
-        let next_msg_id = std::sync::atomic::AtomicU16::new(0x8000); // hub message IDs start high
-        tokio::spawn(async move {
-            let mut interval = tokio::time::interval(std::time::Duration::from_secs(
-                HEARTBEAT_CHECK_INTERVAL_SECS,
-            ));
-            loop {
-                interval.tick().await;
-                let now = now_secs();
-                // Snapshot idle clients
-                let idle_clients: Vec<(Vmac, Arc<Mutex<WsSink>>)> = {
-                    let map = clients_for_hb.lock().await;
-                    map.iter()
-                        .filter(|(_, c)| {
-                            let last = c.last_activity.load(std::sync::atomic::Ordering::Acquire);
-                            now.saturating_sub(last) > HEARTBEAT_IDLE_THRESHOLD_SECS
-                        })
-                        .map(|(vmac, c)| (*vmac, Arc::clone(&c.sink)))
-                        .collect()
-                };
-                for (vmac, sink) in idle_clients {
-                    let msg_id = next_msg_id.fetch_add(1, std::sync::atomic::Ordering::Relaxed);
-                    let hb = ScMessage {
-                        function: ScFunction::HeartbeatRequest,
-                        message_id: msg_id,
-                        originating_vmac: None,
-                        destination_vmac: None,
-                        dest_options: Vec::new(),
-                        data_options: Vec::new(),
-                        payload: Bytes::new(),
-                    };
-                    let mut buf = BytesMut::new();
-                    encode_sc_message(&mut buf, &hb);
-                    let result = tokio::time::timeout(std::time::Duration::from_secs(5), async {
-                        let mut w = sink.lock().await;
-                        w.send(Message::Binary(buf.to_vec().into())).await
-                    })
-                    .await;
-                    if let Err(_) | Ok(Err(_)) = result {
-                        warn!("Hub: heartbeat send failed for {vmac:02x?}, removing client");
-                        let mut map = clients_for_hb.lock().await;
-                        if map
-                            .get(&vmac)
-                            .is_some_and(|client| Arc::ptr_eq(&client.sink, &sink))
-                        {
-                            map.remove(&vmac);
-                        }
-                    }
-                }
-            }
-        });
-    }
-
-    loop {
-        let (tcp_stream, peer_addr) = match listener.accept().await {
-            Ok(v) => v,
-            Err(e) => {
-                warn!("Hub accept error: {e}");
-                continue;
-            }
-        };
-
-        // Reject if too many pre-handshake connections
-        let current = active_connections.load(std::sync::atomic::Ordering::Relaxed);
-        if current >= MAX_ACTIVE_CONNECTIONS {
-            warn!("Hub: rejecting connection from {peer_addr} — max active connections ({MAX_ACTIVE_CONNECTIONS}) reached");
-            drop(tcp_stream);
-            continue;
-        }
-        active_connections.fetch_add(1, std::sync::atomic::Ordering::Relaxed);
-
-        debug!("Hub: new TCP connection from {peer_addr}");
-
-        let acceptor = tls_acceptor.clone();
-        let clients = clients.clone();
-        let conn_counter = active_connections.clone();
-
-        tokio::spawn(async move {
-            // Decrement connection counter when this task exits (any path).
-            struct ConnGuard(std::sync::Arc<std::sync::atomic::AtomicUsize>);
-            impl Drop for ConnGuard {
-                fn drop(&mut self) {
-                    self.0.fetch_sub(1, std::sync::atomic::Ordering::Relaxed);
-                }
-            }
-            let _guard = ConnGuard(conn_counter);
-            // TLS handshake
-            let tls_stream = match acceptor.accept(tcp_stream).await {
-                Ok(s) => s,
-                Err(e) => {
-                    warn!("Hub TLS handshake failed for {peer_addr}: {e}");
-                    return;
-                }
-            };
-
-            // WebSocket upgrade — require and echo the BACnet/SC hub subprotocol.
-            let ws_stream = match tokio_tungstenite::accept_hdr_async(
-                tls_stream,
-                |request: &tokio_tungstenite::tungstenite::handshake::server::Request,
-                 mut response: tokio_tungstenite::tungstenite::handshake::server::Response|
-                 -> Result<
-                    tokio_tungstenite::tungstenite::handshake::server::Response,
-                    tokio_tungstenite::tungstenite::handshake::server::ErrorResponse,
-                > {
-                    if !offers_websocket_subprotocol(request, BACNET_SC_HUB_SUBPROTOCOL) {
-                        return Err(websocket_subprotocol_error_response());
-                    }
-                    response.headers_mut().insert(
-                        "Sec-WebSocket-Protocol",
-                        BACNET_SC_HUB_SUBPROTOCOL.parse().unwrap(),
-                    );
-                    Ok(response)
-                },
-            )
-            .await
-            {
-                Ok(ws) => ws,
-                Err(e) => {
-                    warn!("Hub WebSocket upgrade failed for {peer_addr}: {e}");
-                    return;
-                }
-            };
-
-            let (write, read) = ws_stream.split();
-            let write = Arc::new(Mutex::new(write));
-
-            handle_client(peer_addr, hub_vmac, hub_uuid, read, write, clients).await;
-        });
     }
 }
 
@@ -842,155 +689,6 @@ async fn handle_client(
             debug!("Hub: client {peer_addr} (vmac={vmac:02x?}) disconnected after replacement");
         }
     }
-}
-
-fn offers_websocket_subprotocol(
-    request: &tokio_tungstenite::tungstenite::handshake::server::Request,
-    expected: &str,
-) -> bool {
-    request
-        .headers()
-        .get("Sec-WebSocket-Protocol")
-        .and_then(|v| v.to_str().ok())
-        .map(|s| s.split(',').any(|p| p.trim() == expected))
-        .unwrap_or(false)
-}
-
-fn websocket_subprotocol_error_response(
-) -> tokio_tungstenite::tungstenite::handshake::server::ErrorResponse {
-    tokio_tungstenite::tungstenite::http::Response::builder()
-        .status(tokio_tungstenite::tungstenite::http::StatusCode::BAD_REQUEST)
-        .body(Some(format!(
-            "BACnet/SC hub requires WebSocket subprotocol {BACNET_SC_HUB_SUBPROTOCOL}"
-        )))
-        .expect("static WebSocket error response is valid")
-}
-
-fn unexpected_bvlc_function_error_code(function: ScFunction) -> ErrorCode {
-    match function {
-        ScFunction::Unknown(_) => ErrorCode::BVLC_FUNCTION_UNKNOWN,
-        _ => ErrorCode::UNEXPECTED_DATA,
-    }
-}
-
-fn connect_request_vmac_disposition(vmac: Vmac, hub_vmac: Vmac) -> ConnectRequestVmacDisposition {
-    if vmac == crate::sc_frame::UNKNOWN_VMAC || vmac == BROADCAST_VMAC {
-        ConnectRequestVmacDisposition::CloseReserved
-    } else if vmac == hub_vmac {
-        ConnectRequestVmacDisposition::Nak(
-            ErrorClass::COMMUNICATION,
-            ErrorCode::NODE_DUPLICATE_VMAC,
-        )
-    } else {
-        ConnectRequestVmacDisposition::Accept
-    }
-}
-
-fn hub_client_registration_decision<I>(
-    requested_vmac: Vmac,
-    device_uuid: DeviceUuid,
-    existing_clients: I,
-    max_clients: usize,
-) -> HubClientRegistrationDecision
-where
-    I: IntoIterator<Item = (Vmac, DeviceUuid)>,
-{
-    let mut count = 0usize;
-    let mut requested_vmac_owner = None;
-    let mut existing_uuid_vmac = None;
-
-    for (existing_vmac, existing_uuid) in existing_clients {
-        count += 1;
-        if existing_vmac == requested_vmac {
-            requested_vmac_owner = Some(existing_uuid);
-        }
-        if existing_uuid == device_uuid {
-            existing_uuid_vmac = Some(existing_vmac);
-        }
-    }
-
-    if requested_vmac_owner.is_some_and(|owner_uuid| owner_uuid != device_uuid) {
-        return HubClientRegistrationDecision::NakDuplicateVmac;
-    }
-
-    if let Some(old_vmac) = existing_uuid_vmac {
-        return HubClientRegistrationDecision::Replace { old_vmac };
-    }
-
-    if count >= max_clients {
-        HubClientRegistrationDecision::NakMaxClients
-    } else {
-        HubClientRegistrationDecision::Accept
-    }
-}
-
-async fn registered_client_matches_sink(
-    clients: &Clients,
-    registered_vmac: Vmac,
-    sink: &Arc<Mutex<WsSink>>,
-) -> bool {
-    let map = clients.lock().await;
-    registered_client_matches_sink_in_map(&map, registered_vmac, sink)
-}
-
-fn registered_client_matches_sink_in_map(
-    map: &HashMap<Vmac, HubClient>,
-    registered_vmac: Vmac,
-    sink: &Arc<Mutex<WsSink>>,
-) -> bool {
-    map.get(&registered_vmac)
-        .is_some_and(|client| Arc::ptr_eq(&client.sink, sink))
-}
-
-fn relay_limit_decision(
-    npdu_len: usize,
-    encoded_bvlc_len: usize,
-    max_npdu: u16,
-    max_bvlc: u16,
-) -> RelayLimitDecision {
-    if npdu_len > max_npdu as usize {
-        RelayLimitDecision::DropMaxNpdu
-    } else if encoded_bvlc_len > max_bvlc as usize {
-        RelayLimitDecision::DropMaxBvlc
-    } else {
-        RelayLimitDecision::Send
-    }
-}
-
-/// Build a BVLC-Result NAK message.
-fn build_bvlc_result_nak(
-    message_id: u16,
-    result_for: ScFunction,
-    error_class: ErrorClass,
-    error_code: ErrorCode,
-) -> ScMessage {
-    let error_class = error_class.to_raw().to_be_bytes();
-    let error_code = error_code.to_raw().to_be_bytes();
-    ScMessage {
-        function: ScFunction::Result,
-        message_id,
-        originating_vmac: None,
-        destination_vmac: None,
-        dest_options: Vec::new(),
-        data_options: Vec::new(),
-        payload: Bytes::from(vec![
-            result_for.to_raw(),
-            0x01, // NAK
-            0x00, // error header marker
-            error_class[0],
-            error_class[1],
-            error_code[0],
-            error_code[1],
-        ]),
-    }
-}
-
-/// Current time in seconds since UNIX epoch.
-fn now_secs() -> u64 {
-    std::time::SystemTime::now()
-        .duration_since(std::time::UNIX_EPOCH)
-        .unwrap_or_default()
-        .as_secs()
 }
 
 #[cfg(test)]
